fixed alot of issues with getting and setting values

+++ src/main/java/ie/atu/GrpPoppies/CarPart/CarPart.java
@@ -1,7 +1,7 @@
 package ie.atu.GrpPoppies.CarPart;
 
 public class CarPart {
-        private static int partNumber;
+        private static double partNumber;
         private static String name;
         private static String manufacturer;
         private static String supplier;
@@ -10,10 +10,11 @@ public class CarPart {
         private static String warranty;
         private static String description;
 
+
     public CarPart() {
     }
 
-    public CarPart(int partNumber, String name, String manufacturer, String supplier, int quantity, double price, String warranty, String description) {
+    public CarPart(double partNumber, String name, String manufacturer, String supplier, int quantity, double price, String warranty, String description) {
         this.partNumber = partNumber;
         this.name = name;
         this.manufacturer = manufacturer;
@@ -24,11 +25,11 @@ public class CarPart {
         this.description = description;
     }
 
-    public int getPartNumber() {
+    public double getPartNumber() {
         return partNumber;
     }
 
-    public void setPartNumber(int partNumber) {
+    public void setPartNumber(double partNumber) {
         this.partNumber = partNumber;
     }
 
@@ -87,6 +88,4 @@ public class CarPart {
     public void setDescription(String description) {
         this.description = description;
     }
-
-
 }
